# wire: only borrow/return binaryFreeList buffers at the message level