--- a/wire/common.go
+++ b/wire/common.go
@@ -474,19 +474,22 @@ func writeElements(w io.Writer, elements ...interface{}) error {
 
 // ReadVarInt reads a variable length integer from r and returns it as a uint64.
 func ReadVarInt(r io.Reader, pver uint32) (uint64, error) {
-	discriminant, err := binarySerializer.Uint8(r)
-	if err != nil {
+	buf := binarySerializer.Borrow()
+	if _, err := io.ReadFull(r, buf[:1]); err != nil {
+		binarySerializer.Return(buf)
 		return 0, err
 	}
+	discriminant := buf[0]
 
 	var rv uint64
 	switch discriminant {
 	case 0xff:
-		sv, err := binarySerializer.Uint64(r, littleEndian)
-		if err != nil {
+		if _, err := io.ReadFull(r, buf); err != nil {
+			binarySerializer.Return(buf)
 			return 0, err
 		}
-		rv = sv
+		rv = littleEndian.Uint64(buf)
+		binarySerializer.Return(buf)
 
 		// The encoding is not canonical if the value could have been
 		// encoded using fewer bytes.
@@ -497,11 +500,12 @@ func ReadVarInt(r io.Reader, pver uint32) (uint64, error) {
 		}
 
 	case 0xfe:
-		sv, err := binarySerializer.Uint32(r, littleEndian)
-		if err != nil {
+		if _, err := io.ReadFull(r, buf[:4]); err != nil {
+			binarySerializer.Return(buf)
 			return 0, err
 		}
-		rv = uint64(sv)
+		rv = uint64(littleEndian.Uint32(buf[:4]))
+		binarySerializer.Return(buf)
 
 		// The encoding is not canonical if the value could have been
 		// encoded using fewer bytes.
@@ -512,11 +516,12 @@ func ReadVarInt(r io.Reader, pver uint32) (uint64, error) {
 		}
 
 	case 0xfd:
-		sv, err := binarySerializer.Uint16(r, littleEndian)
-		if err != nil {
+		if _, err := io.ReadFull(r, buf[:2]); err != nil {
+			binarySerializer.Return(buf)
 			return 0, err
 		}
-		rv = uint64(sv)
+		rv = uint64(littleEndian.Uint16(buf[:2]))
+		binarySerializer.Return(buf)
 
 		// The encoding is not canonical if the value could have been
 		// encoded using fewer bytes.
@@ -528,6 +533,7 @@ func ReadVarInt(r io.Reader, pver uint32) (uint64, error) {
 
 	default:
 		rv = uint64(discriminant)
+		binarySerializer.Return(buf)
 	}
 
 	return rv, nil
@@ -536,31 +542,40 @@ func ReadVarInt(r io.Reader, pver uint32) (uint64, error) {
 // WriteVarInt serializes val to w using a variable number of bytes depending
 // on its value.
 func WriteVarInt(w io.Writer, pver uint32, val uint64) error {
-	if val < 0xfd {
-		return binarySerializer.PutUint8(w, uint8(val))
-	}
+	buf := binarySerializer.Borrow()
+	switch {
+	case val < 0xfd:
+		buf[0] = uint8(val)
+		_, err := w.Write(buf[:1])
+		binarySerializer.Return(buf)
+		return err
 
-	if val <= math.MaxUint16 {
-		err := binarySerializer.PutUint8(w, 0xfd)
-		if err != nil {
+	case val <= math.MaxUint16:
+		buf[0] = 0xfd
+		littleEndian.PutUint16(buf[1:3], uint16(val))
+		_, err := w.Write(buf[:3])
+		binarySerializer.Return(buf)
+		return err
+
+	case val <= math.MaxUint32:
+		buf[0] = 0xfe
+		littleEndian.PutUint32(buf[1:5], uint32(val))
+		_, err := w.Write(buf[:5])
+		binarySerializer.Return(buf)
+		return err
+
+	default:
+		buf[0] = 0xff
+		if _, err := w.Write(buf[:1]); err != nil {
+			binarySerializer.Return(buf)
 			return err
 		}
-		return binarySerializer.PutUint16(w, littleEndian, uint16(val))
-	}
 
-	if val <= math.MaxUint32 {
-		err := binarySerializer.PutUint8(w, 0xfe)
-		if err != nil {
-			return err
-		}
-		return binarySerializer.PutUint32(w, littleEndian, uint32(val))
-	}
-
-	err := binarySerializer.PutUint8(w, 0xff)
-	if err != nil {
+		littleEndian.PutUint64(buf, val)
+		_, err := w.Write(buf)
+		binarySerializer.Return(buf)
 		return err
 	}
-	return binarySerializer.PutUint64(w, littleEndian, val)
 }
 
 // VarIntSerializeSize returns the number of bytes it would take to serialize