… wire/bench_test.go
… wire/invvect_test.go
… wire/msgcfheaders.go
… wire/msgcfilter.go
… wire/msggetblocks.go
… wire/msggetcfcheckpt.go
… wire/msggetcfheaders.go
… wire/msggetcfilters.go
… wire/msggetdata.go
… wire/msggetheaders.go
… wire/msgheaders.go
… wire/msginv.go
… wire/msgmerkleblock.go
… wire/msgnotfound.go
… wire/msgping.go
… wire/msgpong.go
… wire/msgreject.go
… wire/netaddress.go
… wire/testdata/block-0000000000000000001602407ac49862a7bca9d00f7f402db20b7be2f5de59d2.blk
… wire/testdata/block-00000000000000000021868c2cefc52a480d173c849412fe81c4e5ab806f94ab.blk