@cfromknecht 2021-02-03
FWIW I rebased this PR (locally) over #1684 and reran the `CalcSighHash` benchmarks.
Since calculating the legacy sighash involves [serializing a copy of the original transaction](https://github.com/btcsuite/btcd/blob/master/txscript/script.go#L658),
optimizing our wire encoding speeds up the sighash calculation speed significantly. 

This is a comparison of CalcSigHash before and after applying the wire optimizations
to this branch: 
```
benchmark                  old ns/op     new ns/op     delta
BenchmarkCalcSigHash-8     3619477       1758789       -51.41%

benchmark                  old allocs     new allocs     delta
BenchmarkCalcSigHash-8     801            712            -11.11%

benchmark                  old bytes     new bytes     delta
BenchmarkCalcSigHash-8     1293354       1290504       -0.22%
```

Using an 80% transaction serialization speedup to ballpark, this indicates roughly 2/3 of
the latency incurred by`CalcSigHash` is transaction serialization, and with this change it
becomes 1/4.