@onyb 2021-11-18
@cfromknecht Could you please rebase this PR, since https://github.com/btcsuite/btcd/pull/1769 landed recently?