@Roasbeef 2023-12-16
Rebased with https://github.com/btcsuite/btcd/pull/2073